Extract getStringWidth
