+++ src/document/doc-utils.js
@@ -1,4 +1,5 @@
 "use strict";
+
 const getLast = require("../utils/get-last.js");
 const { literalline, join } = require("./doc-builders.js");
